barangay_concerns/requirements.txt
Django==4.2.7
Pillow==10.1.0
google-generativeai==0.3.2
python-dotenv==1.0.0
